🚨 fix clippy warning

+++ src/tools/stdio.rs
@@ -35,7 +35,7 @@ pub struct WriteNoWhere;
 
 impl Write for WriteNoWhere {
     fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        Ok(buf.len() as usize)
+        Ok(buf.len())
     }
 
     fn flush(&mut self) -> std::io::Result<()> {
